+++ src/DDTrace/Integrations/Mongo/MongoCollectionIntegration.php
@@ -3,7 +3,7 @@
 namespace DDTrace\Integrations\Mongo;
 
 use DDTrace\Span;
-use DDTrace\Tags;
+use DDTrace\Tag;
 use DDTrace\Integrations\Integration;
 
 final class MongoCollectionIntegration extends Integration
@@ -14,13 +14,13 @@ final class MongoCollectionIntegration extends Integration
     {
         // MongoCollection::__construct ( MongoDB $db , string $name )
         self::traceMethod('__construct', function (Span $span, array $args) {
-            $span->setTag(Tags\MONGODB_DATABASE, (string) $args[0]);
-            $span->setTag(Tags\MONGODB_COLLECTION, $args[1]);
+            $span->setTag(Tag::MONGODB_DATABASE, (string) $args[0]);
+            $span->setTag(Tag::MONGODB_COLLECTION, $args[1]);
         });
         // int MongoCollection::count ([ array $query = array() [, array $options = array() ]] )
         self::traceMethod('count', function (Span $span, array $args) {
             if (isset($args[0])) {
-                $span->setTag(Tags\MONGODB_QUERY, json_encode($args[0]));
+                $span->setTag(Tag::MONGODB_QUERY, json_encode($args[0]));
             }
         });
         // array MongoCollection::createDBRef ( mixed $document_or_id )
@@ -29,58 +29,58 @@ final class MongoCollectionIntegration extends Integration
                 return;
             }
             if (isset($ref['$id'])) {
-                $span->setTag(Tags\MONGODB_BSON_ID, (string) $ref['$id']);
+                $span->setTag(Tag::MONGODB_BSON_ID, (string) $ref['$id']);
             }
             if (isset($ref['$ref'])) {
-                $span->setTag(Tags\MONGODB_COLLECTION, $ref['$ref']);
+                $span->setTag(Tag::MONGODB_COLLECTION, $ref['$ref']);
             }
         });
         // array MongoCollection::distinct ( string $key [, array $query ] )
         self::traceMethod('distinct', function (Span $span, array $args) {
             if (isset($args[1])) {
-                $span->setTag(Tags\MONGODB_QUERY, json_encode($args[1]));
+                $span->setTag(Tag::MONGODB_QUERY, json_encode($args[1]));
             }
         });
         // MongoCursor MongoCollection::find ([ array $query = array() [, array $fields = array() ]] )
         self::traceMethod('find', function (Span $span, array $args) {
             if (isset($args[0])) {
-                $span->setTag(Tags\MONGODB_QUERY, json_encode($args[0]));
+                $span->setTag(Tag::MONGODB_QUERY, json_encode($args[0]));
             }
         });
         // array MongoCollection::findAndModify ( array $query [, array $update [, array $fields [, array $options ]]] )
         self::traceMethod('findAndModify', function (Span $span, array $args) {
-            $span->setTag(Tags\MONGODB_QUERY, json_encode($args[0]));
+            $span->setTag(Tag::MONGODB_QUERY, json_encode($args[0]));
         });
         // array MongoCollection::findOne ([ array $query = array() [, array $fields = array()
         // [, array $options = array() ]]] )
         self::traceMethod('findOne', function (Span $span, array $args) {
             if (isset($args[0])) {
-                $span->setTag(Tags\MONGODB_QUERY, json_encode($args[0]));
+                $span->setTag(Tag::MONGODB_QUERY, json_encode($args[0]));
             }
         });
         // array MongoCollection::getDBRef ( array $ref )
         self::traceMethod('getDBRef', function (Span $span, array $args) {
             if (isset($args[0]['$id'])) {
-                $span->setTag(Tags\MONGODB_BSON_ID, $args[0]['$id']);
+                $span->setTag(Tag::MONGODB_BSON_ID, $args[0]['$id']);
             }
             if (isset($args[0]['$ref'])) {
-                $span->setTag(Tags\MONGODB_COLLECTION, $args[0]['$ref']);
+                $span->setTag(Tag::MONGODB_COLLECTION, $args[0]['$ref']);
             }
         });
         // bool|array MongoCollection::remove ([ array $criteria = array() [, array $options = array() ]] )
         self::traceMethod('remove', function (Span $span, array $args) {
             if (isset($args[0])) {
-                $span->setTag(Tags\MONGODB_QUERY, json_encode($args[0]));
+                $span->setTag(Tag::MONGODB_QUERY, json_encode($args[0]));
             }
         });
         // bool MongoCollection::setReadPreference ( string $read_preference [, array $tags ] )
         self::traceMethod('setReadPreference', function (Span $span, array $args) {
-            $span->setTag(Tags\MONGODB_READ_PREFERENCE, $args[0]);
+            $span->setTag(Tag::MONGODB_READ_PREFERENCE, $args[0]);
         });
         // bool|array MongoCollection::update ( array $criteria , array $new_object [, array $options = array() ] )
         self::traceMethod('update', function (Span $span, array $args) {
             if (isset($args[0])) {
-                $span->setTag(Tags\MONGODB_QUERY, json_encode($args[0]));
+                $span->setTag(Tag::MONGODB_QUERY, json_encode($args[0]));
             }
         });
         // Methods that don't need extra tags added