+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file - [read more
 
 ## [Unreleased]
 ### Added
+- Migration from namespace based constants to class based constants for tags, formats and types #207
+- Track integration loading to avoid re-loading unnecessary ones #211
 - Documenting release steps #223
 
 ### Fixed
@@ -12,6 +14,7 @@ All notable changes to this project will be documented in this file - [read more
 - PHP code compatibility with PHP 5.4 #194
 - Move framework tests to tests root folder #198
 - Move integrations tests to tests root folder #200
+- Removal of external dependencies to support auto-instrumentation #206
 - Allow testing of multiple library versions #203
 - Downgrade of phpunit to 4.* in order to prepare for php 5.4 #208
 - Configurable autofinishing of unfinished spans on tracer flush #217